Ensure that new users don't make the same mistakes as I did
Before I got the hang of this, I just copied the `-devel` package to the folder, but solbuild was throwing errors and making me get really annoyed. I later figured out that I need both the regular package and the `-devel` for it to work. I hope that this will prevent new packagers from making the same mistake.