Update Package Inclusion Policy.
This update to our package inclusion policy removes the 2-distro waiver, as it was found to enable the inclusion of software which would not have otherwise been accepted.

Furthermore, a clause has been added under Value Add to enable us to be more stringent on accepting or rejecting applications which are designed for another, specific, operating system, or which rely on libraries that are designed with the intent of functioning on a specific operating system.