fixed some things
after messing around with shells, i noticed that you do not run the "chsh" command as root, as it will only change the default shell for root. i updated the examples to reflect this, along with adding an example for fish, wi and a link to herbet xu's website for dash, which also contains a link to the source code on git.kernel.org. i also added a troubleshooting guide in the case of permission issues with chsh due to a shell not being in /etc/shells for fish and zsh.