new package: Update example to subject: verb version
With the monorepo, we use the `worklog.py` script and the GH UI, which
shows the git shortlog at a glance.

Therefore, the git shortlog messages should be as informative as possible
so people don't have to go chasing the individual commit logs for
relevant version information when looking for specific updates or
summarising the weekly development churn in the forum sync notes.

This commit changes the old (and no longer useful) "Initial inclusion"
git shortlog message to a more useful "packagename: Add at version"
git shortlog message.

Note that we already require "packagename: Update to version" for the
same reason.

Signed-off-by: Rune Morling <ermo@serpentos.com>