docs/packaging: Update docs for `appstream` & `appstream-generator` (#619)
We are nuking our use of `appstream-glib` and `appstream-builder`.
Explain how to run `appstreamcli compose` on a single package as well
as running `appstream-generator` against a repository.

The https://github.com/getsolus/solus-appstream-data repo will be
archived so any relevant docs from there will need to be moved over.

The "debugging failures" section is updated to account for the most
common failures and a writing metainfo is also provided on creating
metainfo from scratch.

Resolves #628.