deps, config: Largely redo both the deps and configs
In order to try to fix https://github.com/getsolus/help-center-docs/issues/640, I ended up going over all of the dependencies, as the root cause of all but one of the issues was a mismatch between the `react` version and `@types/react`. There is still one outstanding issue that I've been unable to figure out so far: The IdealImage module can't be found. From scouring issues, I gathered that this is usually caused by missing imports, itself stemming from the configuration files being in JavaScript and not TypeScript. So, I migrated our configs to TypeScript and modernized them, to no avail.

In this whole journey, I discovered that the plugin we were using for search doesn't support Docusaurus v3. That it ever worked at all was just blind luck. Luckily, there is a maintained and updated fork that I switched us to. The UI is better, too, so yay.

Ref https://github.com/getsolus/help-center-docs/issues/640

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>