editions: Split edition definitions to own file
Right now, the data for our website is a bit spread out. People and teams live in `src/data`, but others like our editions and top-level categories live in the component and page files. Let's make everything easier to find. This moves the edition definitions to a file under `src/data`, making the component responsible only for the actual UI of the site.

~~I did this all jumbled up with https://github.com/getsolus/help-center-docs/pull/641, so it might be a good idea to merge this PR *after* that one, just in case I missed something when trying to separate all of my changes.~~

Depends on https://github.com/getsolus/help-center-docs/pull/660

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>